requirements.txt
langchain
langchain-community
langchain-openai
nemoguardrails
python-dotenv
PyYAML
